refactor: move chat slash commands to command layer
